Improve mixin interface names
Also:
 * Add associate rendering actions with multiple message types.
 * Pull `RenderActions` to the package level.
