+++ mc-java-signal/src/main/kotlin/io/spine/tools/mc/java/signal/Renderers.kt
@@ -34,7 +34,7 @@ import io.spine.protodata.java.ClassName
 import io.spine.protodata.renderer.SourceFile
 import io.spine.tools.code.Java
 import io.spine.tools.mc.java.MessageTypeRenderer
-import io.spine.tools.mc.java.WithTypeList
+import io.spine.tools.mc.java.TypeListActions
 import io.spine.tools.mc.java.field.FieldClass
 import io.spine.tools.mc.java.field.superClassName
 import io.spine.tools.mc.java.settings.SignalSettings
@@ -49,16 +49,18 @@ import io.spine.tools.psi.java.execute
 internal abstract class SignalRenderer<V> :
     MessageTypeRenderer<V, SignalSettings>(),
     SignalPluginComponent
-        where V : EntityState<File>, V : WithTypeList {
+        where V : EntityState<File>, V : TypeListActions {
 
     /**
      * The settings for the kind of signals served by this renderer, obtained from [settings].
      */
     protected abstract val typeSettings: Signals
 
+    //TODO:2024-07-29:alexander.yevsyukov: Check `pattern` and `actions` instead.
     override val enabledBySettings: Boolean
         get() = typeSettings.generateFields.hasSuperclass()
 
+    //TODO:2024-07-29:alexander.yevsyukov: Transform to a parameter, avoiding default value in settings.
     private val fieldSupertype: ClassName by lazy {
         typeSettings.generateFields.superClassName
     }
@@ -66,6 +68,10 @@ internal abstract class SignalRenderer<V> :
     @OverridingMethodsMustInvokeSuper
     override fun doRender(type: MessageType, file: SourceFile<Java>) {
         execute {
+            //TODO:2024-07-29:alexander.yevsyukov: Uncomment.
+            // ImplementInterface(type, file, superInterface, context = context!!).run {
+            //    render()
+            // }
             FieldClass(type, file, fieldSupertype, context!!).run {
                 render()
             }
